feat(ui): add HTML block selector window
